Fixed bugs

+++ src/Memory.h
@@ -4,13 +4,14 @@
 #define MEM_SIZE 32
 
 #include <vector>
+#include <stdint.h>
 #include <stdexcept>
 
 class Memory {
 private:
-	std::vector<int> stack_;
-	int* memory_;
-	int* ptr_;
+	std::vector<uint8_t> stack_;
+	uint8_t* memory_;
+	uint8_t* ptr_;
 
 public:
 	Memory();
@@ -21,14 +22,14 @@ public:
 	void incPtr();
 	void decPtr();
 
-	void memWrite(int val);
+	void memWrite(uint8_t val);
 	void copyFromStack();
 	void popFromStack();
 	void pushToStack();
 
 	void reset();
 	int currVal();
-	int* getMem();
-	std::vector<int> getStack();
+	uint8_t* getMem();
+	std::vector<uint8_t> getStack();
 };
 #endif // !MEMORY_H
